requirements.txt
streamlit>=1.36
pyyaml>=6.0
numpy>=1.24
pandas>=2.0
scipy>=1.10
altair
